backend/requirements.txt
django
djangorestframework
django-cors-headers
psycopg[binary]
python-dotenv
dj-database-url
psycopg
djangorestframework-simplejwt
celery
requests
redis
boto3
openai>=1.0.0
pypdf>=3.0.0
tiktoken